first split of logic
